feat(oidc-client-filter): move filter as pre-step for fixing split packages
* Pre-step for: https://github.com/quarkusio/quarkus/issues/51872
* After some time, we will remove the filter class, thus fixing the split packages
* I am going to create Quarkus Update recipe to migrate users applications automatically
