+++ extensions/oidc-client-filter/deployment/src/test/java/io/quarkus/oidc/client/filter/ExtendedOidcClientRequestFilter.java
@@ -3,6 +3,8 @@ package io.quarkus.oidc.client.filter;
 import jakarta.annotation.Priority;
 import jakarta.ws.rs.Priorities;
 
+import io.quarkus.oidc.client.resteasy.filter.OidcClientRequestFilter;
+
 @Priority(Priorities.AUTHENTICATION)
 public class ExtendedOidcClientRequestFilter extends OidcClientRequestFilter {
 }
