+++ extensions/oidc-client-filter/deployment/src/test/java/io/quarkus/oidc/client/filter/ProtectedResourceServiceCustomProviderConfigPropOidcClient.java
@@ -6,6 +6,8 @@ import jakarta.ws.rs.Path;
 import org.eclipse.microprofile.rest.client.annotation.RegisterProvider;
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
+import io.quarkus.oidc.client.resteasy.filter.OidcClientRequestFilter;
+
 @RegisterProvider(OidcClientRequestFilter.class)
 @RegisterRestClient
 @Path("/")
