+++ extensions/oidc-client-filter/runtime/src/main/java/io/quarkus/oidc/client/resteasy/filter/OidcClientRequestFilter.java
@@ -1,4 +1,4 @@
-package io.quarkus.oidc.client.filter;
+package io.quarkus.oidc.client.resteasy.filter;
 
 import java.util.Optional;
 
@@ -11,10 +11,6 @@ import jakarta.ws.rs.ext.Provider;
 import io.quarkus.oidc.client.filter.runtime.AbstractOidcClientRequestFilter;
 import io.quarkus.oidc.client.filter.runtime.OidcClientFilterConfig;
 
-/**
- * @deprecated use the {@link io.quarkus.oidc.client.resteasy.filter.OidcClientRequestFilter} filter instead
- */
-@Deprecated(since = "3.31", forRemoval = true)
 @Provider
 @Singleton
 @Priority(Priorities.AUTHENTICATION)
